+++ src/vite-env.d.ts
@@ -25,6 +25,7 @@ declare module 'ionicons/icons' {
   export const chevronForwardOutline: string;
   export const imageOutline: string;
   export const swapHorizontalOutline: string;
+  export const albumsOutline: string;
   export const homeOutline: string;
   export const scaleOutline: string;
   export const bedOutline: string;